+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/exception/Exception.kt
@@ -1,0 +1,8 @@
+package com.github.sam0delkin.intellijpsa.exception
+
+import java.lang.RuntimeException
+
+class Exception {
+    class IndexNotReadyException : RuntimeException()
+    class IndexingDisabledException : RuntimeException()
+}