+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/index/PsiFileProcessor.kt
@@ -7,14 +7,10 @@ import com.intellij.psi.search.PsiElementProcessor
 import com.intellij.psi.util.elementType
 import com.intellij.refactoring.suggested.startOffset
 import com.jetbrains.rd.util.string.printToString
-import java.io.File
 
 class PsiFileProcessor(
     private val completionService: CompletionService,
     private val notIndexedElements: ArrayList<String>,
-    private val completionResultFilePaths: HashMap<String, String>,
-    private val goToResultFilePaths: HashMap<String, String>,
-    private val resultMap: HashMap<String, String>
 ) :
     PsiElementProcessor<PsiElement> {
     override fun execute(currentElement: PsiElement): Boolean {
@@ -25,46 +21,6 @@ class PsiFileProcessor(
 
         val path = PsiUtils.getPsiElementLabel(currentElement)
         if (!settings.elementPaths.contains(path)) {
-            return true
-        }
-
-        var goToResult: String?
-        var completionResult: String?
-        try {
-            val goToKey = PsaIndex.generateGoToKey(currentElement)
-            goToResult = this.goToResultFilePaths.get(goToKey)
-
-            if (null !== goToResult) {
-                val file = File(goToResult)
-                goToResult = file.readText()
-
-                resultMap.set(goToKey, goToResult)
-
-                file.delete()
-                this.goToResultFilePaths.remove(goToKey)
-            }
-        } catch (_: Exception) {
-            return true
-        }
-
-        try {
-            val completionsKey = PsaIndex.generateCompletionsKey(currentElement)
-            completionResult = this.completionResultFilePaths.get(completionsKey)
-
-            if (null !== completionResult) {
-                val file = File(completionResult)
-                completionResult = file.readText()
-
-                resultMap.set(completionsKey, completionResult)
-
-                file.delete()
-                this.completionResultFilePaths.remove(completionsKey)
-            }
-        } catch (_: Exception) {
-            return true
-        }
-
-        if (null !== goToResult || null !== completionResult) {
             return true
         }